Change the return type of close_dependencies.
For solving dependencies, we don't want Provides to be a Proposition.
It isn't in the first place.  Instead, what we want is a dictionary that
is keyed on something that is provided and has values of everything that
provides that thing.

We can then take that dictionary and use it against all the
Propositions:  everywhere we see something that is required, we look it
up in the dictionary.  If it's in there, we substitute in all the
providers.  If there's more than one provider, we'll make it a boolean
expression like "A or B or C".