Add unit propagation, the beginnings of dependency solving.
This is an optimization step for 3SAT - Basically if you ever see an
expression that consists of just a package name, or not-package name,
you can reduce it to true or false.  You know whether that package can
be installed or not without doing anything else.  Then you run through
all the other expressions and remove any references to those packages.

Continue this process until you're through making changes.  The result
is a dictionary of assignments (package -> true/false) and a modified
set of expressions.

While it is not expected that unit propagation will completely solve
things, for our extremely simple cases of just "X requires Y" or "X and
not Y", it does so.  For more complicated stuff, there's still a ways to
go.  Version numbers in particular are causing problems.