Remove some duplicated groups
If a group is provided in a previous AND clause it doesn't need to be
included in future AND clauses. Additionally, req_providers doesn't need
to add the groupid= atom, since depclose_package already does that, and
it's possible for req_providers to return nothing, in case requirements
are all satisfied via parent expressions.