Remove the Eq and PartialEq traits from DepExpression.
The derived implementation of PartialEq is not correct for
DepExpression, since the order within the And and Or vectors should not
matter for expression equality. Comparing two expressions for equality
isn't actually something that's needed, so remove the traits to avoid
confusion. Replace the one part where == was being used with an
explicit match.