Simplify dependency expressions.
The DepExpression atoms do not need Requirement data. The only variables
that make sense to be turned on or off are packages (groups), so the
only thing going into depsolve should be group ids. The problem that
requirements as part of the expression was trying to solve should
actually be solved within depclose: the returned expression for a
particular group id needs to include conflict and obsolete data, also in
terms of group ids.

Getting complicated crud out of the DepExression means we can also get
the weirdo pointer types out of the depexpression. The returned
expression for complicated packages can still be pretty big, but not
*that* big, so we don't need Rc, which removes the need for RefCell, and
getting rid of those is a huge plus: DepExpression can no longer
contain cycles or leak memory of have runtime borrow problems.

The last need for a pointer was in DepExpression::Not. To be able to do
Not(expr), the inner expression needed to be in a Box, or something.
But, 1) we never actually used Not like that, and 2) Not(A (And|Or) B)
can be changed into combinations of Not(id) expression via De Morgan's
laws. So no more weirdo pointers.